+++ api/instances/workbench-adapters.ts
@@ -1,0 +1,5 @@
+import { garbageLocalWorkbenchAdapter } from './garbage/workbench.js'
+
+export const registeredLocalWorkbenchAdapters = [
+  garbageLocalWorkbenchAdapter,
+]